Fix crate name inconsistencies
If a crate as a different name from its library (see example below) the
docs are generated into a folder with the name of the target. Hence
instead of using the package name, we should use the target name.

```toml
[package]
name = "my-cratename"

[lib]
name = "my-actual-cratename"
```

Currently it selects the first target, which means that it might select
a binary as the target. Normally on only generates docs for a library,
so instead of choosing the first target we could change it to choose the
first library target and then fallback to a binary target if there are
no library targets.

Related to #4